Move tensor attribute to search::tensor namespace.
