+++ searchlib/src/vespa/searchlib/features/attributefeature.cpp
@@ -24,10 +24,10 @@ using search::attribute::IAttributeVector;
 using search::attribute::BasicType;
 using search::attribute::CollectionType;
 using search::attribute::ConstCharContent;
-using search::attribute::DenseTensorAttribute;
+using search::tensor::DenseTensorAttribute;
 using search::attribute::IntegerContent;
 using search::attribute::FloatContent;
-using search::attribute::TensorAttribute;
+using search::tensor::TensorAttribute;
 using search::attribute::WeightedConstCharContent;
 using search::attribute::WeightedIntegerContent;
 using search::attribute::WeightedFloatContent;
