+++ searchlib/src/vespa/searchlib/features/dense_tensor_attribute_executor.h
@@ -7,7 +7,7 @@
 #include <vespa/vespalib/tensor/dense/mutable_dense_tensor_view.h>
 
 namespace search {
-namespace attribute { class DenseTensorAttribute; }
+namespace tensor { class DenseTensorAttribute; }
 namespace features {
 
 /**
@@ -17,12 +17,12 @@ namespace features {
 class DenseTensorAttributeExecutor : public fef::FeatureExecutor
 {
 private:
-    const search::attribute::DenseTensorAttribute *_attribute;
+    const search::tensor::DenseTensorAttribute *_attribute;
     vespalib::tensor::MutableDenseTensorView *_tensorView;
     vespalib::eval::TensorValue _tensor;
 
 public:
-    DenseTensorAttributeExecutor(const search::attribute::DenseTensorAttribute *attribute);
+    DenseTensorAttributeExecutor(const search::tensor::DenseTensorAttribute *attribute);
     virtual void execute(fef::MatchData &data);
 };
