+++ searchlib/src/vespa/searchlib/tensor/generic_tensor_store.cpp
@@ -17,7 +17,7 @@ using vespalib::tensor::TypedBinaryFormat;
 
 namespace search {
 
-namespace attribute {
+namespace tensor {
 
 constexpr size_t MIN_BUFFER_CLUSTERS = 1024;
 
@@ -120,6 +120,6 @@ GenericTensorStore::setTensor(const Tensor &tensor)
     return raw.ref;
 }
 
-}  // namespace search::attribute
+}  // namespace search::tensor
 
 }  // namespace search
