+++ searchlib/src/vespa/searchlib/tensor/tensor_store.h
@@ -10,7 +10,7 @@ namespace vespalib { namespace tensor { class Tensor; } }
 
 namespace search {
 
-namespace attribute {
+namespace tensor {
 
 /**
  * Class for storing serialized tensors in memory, used by TensorAttribute.
@@ -76,6 +76,6 @@ public:
 };
 
 
-}  // namespace search::attribute
+}  // namespace search::tensor
 
 }  // namespace search
